organize config lib
